cv_backend/requirements.txt
Flask==3.0.3
keras==3.3.3
matplotlib==3.8.4
numpy==1.26.4
oauthlib==3.2.2
opencv_python==4.9.0.80
Pillow==10.3.0
requests_oauthlib==2.0.0
